split out BaseSettings from Detector.h
